Refactor protocol adapters to support non spring based implementations
