+++ adapters/amqp-vertx/pom.xml
@@ -16,7 +16,8 @@
   <modelVersion>4.0.0</modelVersion>
   <parent>
     <groupId>org.eclipse.hono</groupId>
-    <artifactId>hono-adapters</artifactId>
+    <artifactId>hono-adapter-base-spring</artifactId>
+    <relativePath>../base-spring</relativePath>
     <version>1.4.0-SNAPSHOT</version>
   </parent>
   <artifactId>hono-adapter-amqp-vertx</artifactId>
