+++ adapters/coap-vertx-base/src/main/java/org/eclipse/hono/adapter/coap/impl/VertxBasedCoapAdapter.java
