+++ adapters/pom.xml
@@ -36,6 +36,7 @@
     <module>mqtt-vertx</module>
     <module>mqtt-vertx-base</module>
     <module>sigfox-vertx</module>
+    <module>base-spring</module>
   </modules>
 
   <name>Hono Protocol Adapters</name>
@@ -65,18 +66,6 @@
     <dependency>
       <groupId>org.eclipse.hono</groupId>
       <artifactId>hono-service-base</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-autoconfigure</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-starter-logging</artifactId>
     </dependency>
     <dependency>
       <groupId>org.yaml</groupId>
