+++ adapters/sigfox-vertx/pom.xml
@@ -14,9 +14,10 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>org.eclipse.hono</groupId>
-        <artifactId>hono-adapters</artifactId>
-        <version>1.4.0-SNAPSHOT</version>
+      <groupId>org.eclipse.hono</groupId>
+      <artifactId>hono-adapter-base-spring</artifactId>
+      <relativePath>../base-spring</relativePath>
+      <version>1.4.0-SNAPSHOT</version>
     </parent>
     <artifactId>hono-adapter-sigfox-vertx</artifactId>
     <name>Hono Sigfox Adapter</name>
